Warning fixes

+++ src/app/rhi/rhi_shader.cpp
@@ -23,13 +23,13 @@ VkResult shader_pipeline::build(const VkDevice device)
 	push_constant_range.stageFlags = VK_SHADER_STAGE_ALL;
 	for (VkShaderCreateInfoEXT& create_info : shaders_create_info_)
 	{
-		create_info.setLayoutCount = layouts.size();
+		create_info.setLayoutCount = static_cast<uint32_t>(layouts.size());
 		create_info.pSetLayouts = layouts.data();
 		create_info.pushConstantRangeCount = 1;
 		create_info.pPushConstantRanges = &push_constant_range;
 	}
 	shaders.resize(shaders_create_info_.size());
-	if (const VkResult result = vkCreateShadersEXT(device, shaders_create_info_.size(), shaders_create_info_.data(), nullptr, shaders.data()); result != VK_SUCCESS) return result;
+	if (const VkResult result = vkCreateShadersEXT(device, static_cast<uint32_t>(shaders_create_info_.size()), shaders_create_info_.data(), nullptr, shaders.data()); result != VK_SUCCESS) return result;
 	{
 		const auto obj_name = std::format("{}_vert", name);
 		const VkDebugUtilsObjectNameInfoEXT object_name = rhi_helpers::add_name(VK_OBJECT_TYPE_SHADER_EXT, reinterpret_cast<uint64_t>(shaders.data()[0]), obj_name.c_str());
@@ -41,7 +41,7 @@ VkResult shader_pipeline::build(const VkDevice device)
 		if (const VkResult result = vkSetDebugUtilsObjectNameEXT(device, &frag_name); result != VK_SUCCESS) return result;
 	}
 	{
-		const VkPipelineLayoutCreateInfo pl_info = rhi_helpers::create_pipeline_layout_create_info(push_constant_range, 1, layouts.data(), layouts.size());
+		const VkPipelineLayoutCreateInfo pl_info = rhi_helpers::create_pipeline_layout_create_info(push_constant_range, 1, layouts.data(), static_cast<uint32_t>(layouts.size()));
 		VkPipelineLayout layout{};
 		if (const VkResult result = vkCreatePipelineLayout(device, &pl_info, nullptr, &layout); result != VK_SUCCESS) return result;
 		pipeline_layout = layout;
